Refactor fields/lanes
