+++ modules/ui/preset.js
@@ -23,7 +23,6 @@ export function uiPreset(context) {
     // Field Constructor
     function UIField(field, entity, show) {
         field = _.clone(field);
-        console.log(field);
         field.input = uiFields[field.type](field, context)
             .on('change', function(t, onInput) {
                 dispatch.call('change', field, t, onInput);
